first pass on pondering agent
