+++ agent/zobrist/__init__.py
@@ -1,0 +1,5 @@
+from agent.zobrist.hashing import create_board_hash, update_board_hash
+
+class Zobrist:
+    create_board_hash = create_board_hash
+    update_board_hash = update_board_hash
