+++ app.py
@@ -1,4 +1,4 @@
-from ui import App
+from ui.app import App
 
 if __name__ == "__main__":
     app = App()
