+++ ui/model/agent_type.py
@@ -2,11 +2,13 @@ from enum import Enum
 
 from agent.default.agent import DefaultAgent
 from agent.brandon.agent import BrandonAgent
+from agent.brandon.agent_ponder import BrandonPonderer
 
 
 class AgentType(Enum):
     DEFAULT = "Default"
     BRANDON = "2-ply negamax"
+    BRANDON_PONDERER = "Pondering 2-ply negamax"
 
     def create(self):
         """
@@ -15,4 +17,5 @@ class AgentType(Enum):
         return {
             AgentType.DEFAULT: DefaultAgent,
             AgentType.BRANDON: BrandonAgent,
+            AgentType.BRANDON_PONDERER: BrandonPonderer,
         }[self]()
