+++ ui/model/config.py
@@ -17,14 +17,14 @@ class Config:
 
     layout: BoardLayout = DEFAULT_LAYOUT
     move_limit: int = 40
-    player_type_p1: PlayerType = PlayerType.COMPUTER
+    player_type_p1: PlayerType = PlayerType.HUMAN
     player_type_p2: PlayerType = PlayerType.COMPUTER
     time_limit_p1: float = 10.0
     time_limit_p2: float = 10.0
     heuristic_type_p1: HeuristicType = HeuristicType.WEIGHTED_NORMALIZED
     heuristic_type_p2: HeuristicType = HeuristicType.BRANDON
     agent_type_p1: AgentType = AgentType.DEFAULT
-    agent_type_p2: AgentType = AgentType.BRANDON
+    agent_type_p2: AgentType = AgentType.BRANDON_PONDERER
 
     @classmethod
     def from_default(cls):
